refactor: incorporated PR feedback: moved validate methods to workflow classes. Renamed substeps pkg to workflow. Moved step factory to workflow package.
