+++ src/main/java/com/synopsys/integration/jenkins/polaris/workflow/GetPolarisCliResponseContent.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.jenkins.polaris.substeps;
+package com.synopsys.integration.jenkins.polaris.workflow;
 
 import java.io.IOException;
 import java.nio.file.Files;
