+++ src/main/java/com/synopsys/integration/jenkins/polaris/workflow/PolarisWorkflowStepFactory.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.jenkins.polaris.extensions.buildstep;
+package com.synopsys.integration.jenkins.polaris.workflow;
 
 import java.io.IOException;
 
@@ -28,11 +28,6 @@ import com.synopsys.integration.function.ThrowingConsumer;
 import com.synopsys.integration.jenkins.extensions.JenkinsIntLogger;
 import com.synopsys.integration.jenkins.polaris.extensions.global.PolarisGlobalConfig;
 import com.synopsys.integration.jenkins.polaris.extensions.tools.PolarisCli;
-import com.synopsys.integration.jenkins.polaris.substeps.CreatePolarisEnvironment;
-import com.synopsys.integration.jenkins.polaris.substeps.ExecutePolarisCli;
-import com.synopsys.integration.jenkins.polaris.substeps.GetPathToPolarisCli;
-import com.synopsys.integration.jenkins.polaris.substeps.GetPolarisCliResponseContent;
-import com.synopsys.integration.jenkins.polaris.substeps.GetTotalIssueCount;
 import com.synopsys.integration.polaris.common.configuration.PolarisServerConfig;
 import com.synopsys.integration.polaris.common.service.PolarisService;
 import com.synopsys.integration.polaris.common.service.PolarisServicesFactory;
